requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 24.11.0
accelerate=1.2.0=pypi_0
aiohappyeyeballs=2.4.3=py39haa95532_0
aiohttp=3.10.5=py39h827c3e9_0
aiosignal=1.2.0=pyhd3eb1b0_0
arrow-cpp=16.1.0=h7cd61ee_0
async-timeout=4.0.3=py39haa95532_0
attrs=24.2.0=py39haa95532_0
aws-c-auth=0.6.19=h2bbff1b_0
aws-c-cal=0.5.20=h2bbff1b_0
aws-c-common=0.8.5=h2bbff1b_0
aws-c-compression=0.2.16=h2bbff1b_0
aws-c-event-stream=0.2.15=hd77b12b_0
aws-c-http=0.6.25=h2bbff1b_0
aws-c-io=0.13.10=h2bbff1b_0
aws-c-mqtt=0.7.13=h2bbff1b_0
aws-c-s3=0.1.51=h2bbff1b_0
aws-c-sdkutils=0.1.6=h2bbff1b_0
aws-checksums=0.1.13=h2bbff1b_0
aws-crt-cpp=0.18.16=hd77b12b_0
aws-sdk-cpp=1.10.55=hd77b12b_0
beautifulsoup4=4.12.3=pypi_0
blas=1.0=mkl
boost-cpp=1.82.0=h59b6b97_2
bottleneck=1.4.2=py39hc99e966_0
brotli-python=1.0.9=py39hd77b12b_8
bzip2=1.0.8=h2bbff1b_6
c-ares=1.19.1=h2bbff1b_0
ca-certificates=2024.11.26=haa95532_0
certifi=2024.8.30=py39haa95532_0
cffi=1.17.1=py39h827c3e9_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
colorama=0.4.6=py39haa95532_0
cuda-cccl=12.4.127=0
cuda-cudart=11.8.89=0
cuda-cudart-dev=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-libraries-dev=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvrtc-dev=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-profiler-api=12.4.127=0
cuda-runtime=11.8.0=0
cuda-version=11.8=hcce14f8_3
cudatoolkit=11.8.0=hd77b12b_0
datasets=2.19.1=py39haa95532_0
dill=0.3.8=py39haa95532_0
filelock=3.13.1=py39haa95532_0
freetype=2.12.1=ha860e81_0
frozenlist=1.5.0=py39h827c3e9_0
fsspec=2024.3.1=py39haa95532_0
future=1.0.0=py39haa95532_0
gflags=2.2.2=hd77b12b_1
giflib=5.2.2=h7edc060_0
glog=0.5.0=hd77b12b_1
gmpy2=2.1.2=py39h7f96b67_0
huggingface_hub=0.24.6=py39haa95532_0
idna=3.7=py39haa95532_0
intel-openmp=2023.1.0=h59b6b97_46320
jinja2=3.1.4=py39haa95532_1
jpeg=9e=h827c3e9_3
lcms2=2.12=h83e58a3_0
lerc=3.0=hd77b12b_0
libabseil=20240116.2=cxx17_h5da7b33_0
libboost=1.82.0=h3399ecb_2
libbrotlicommon=1.0.9=h2bbff1b_8
libbrotlidec=1.0.9=h2bbff1b_8
libbrotlienc=1.0.9=h2bbff1b_8
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcurand=10.3.5.147=0
libcurand-dev=10.3.5.147=0
libcurl=8.9.1=h0416ee5_0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.17=h2bbff1b_1
libevent=2.1.12=h56d1f94_1
libffi=3.4.4=hd77b12b_1
libgrpc=1.62.2=hf25190f_0
libjpeg-turbo=2.0.0=h196d8e1_0
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvfatbin=12.4.127=0
libnvfatbin-dev=12.4.127=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libpng=1.6.39=h8cc25b3_0
libprotobuf=4.25.3=hf2fb9eb_0
libssh2=1.11.1=h2addb87_0
libthrift=0.15.0=h4364b78_2
libtiff=4.5.1=hd77b12b_0
libuv=1.48.0=h827c3e9_0
libwebp=1.3.2=hbc33d0d_0
libwebp-base=1.3.2=h3d04722_1
lyricsgenius=3.0.1=dev_0
lz4-c=1.9.4=h2bbff1b_1
markupsafe=2.1.3=py39h2bbff1b_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py39h2bbff1b_1
mkl_fft=1.3.11=py39h827c3e9_0
mkl_random=1.2.8=py39hc64d2fc_0
mpc=1.1.0=h7edee0f_1
mpfr=4.0.2=h62dcd97_1
mpir=3.0.0=hec2e145_1
mpmath=1.3.0=py39haa95532_0
multidict=6.1.0=py39h827c3e9_0
multiprocess=0.70.15=py39haa95532_0
networkx=3.2.1=py39haa95532_0
ninja=1.12.1=haa95532_0
ninja-base=1.12.1=h4157e71_0
numexpr=2.10.1=py39h4cd664f_0
numpy=1.26.4=py39h055cbcc_0
numpy-base=1.26.4=py39h65a83cf_0
openjpeg=2.5.2=hae555c5_0
openssl=3.4.0=h2466b09_0
orc=2.0.1=hd8d391b_0
packaging=24.1=py39haa95532_0
pandas=2.2.3=py39h5da7b33_0
pillow=11.0.0=py39hb5480e2_0
pip=24.2=py39haa95532_0
propcache=0.2.0=py39h827c3e9_0
psutil=6.1.0=pypi_0
pyarrow=16.1.0=py39hc64d2fc_0
pycparser=2.21=pyhd3eb1b0_0
pysocks=1.7.1=py39haa95532_0
python=3.9.20=h8205438_1
python-dateutil=2.9.0post0=py39haa95532_2
python-tzdata=2023.3=pyhd3eb1b0_0
python-xxhash=2.0.2=py39h2bbff1b_1
pytorch=2.5.1=py3.9_cuda11.8_cudnn9_0
pytorch-cuda=11.8=h24eeafa_6
pytorch-mutex=1.0=cuda
pytz=2024.1=py39haa95532_0
pyyaml=6.0.2=py39h827c3e9_0
re2=2022.04.01=hd77b12b_0
regex=2024.9.11=py39h827c3e9_0
requests=2.32.3=py39haa95532_1
safetensors=0.4.5=py39haa69e8e_1
setuptools=75.1.0=py39haa95532_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.2.1=hcdb6601_0
soupsieve=2.6=pypi_0
sqlite=3.45.3=h2bbff1b_0
sympy=1.13.1=pypi_0
tbb=2021.8.0=h59b6b97_0
tk=8.6.14=h0416ee5_0
tokenizers=0.21.0=pypi_0
torchaudio=2.5.1=pypi_0
torchvision=0.20.1=pypi_0
tqdm=4.66.5=py39h9909e9c_0
transformers=4.47.0=pypi_0
typing-extensions=4.11.0=py39haa95532_0
typing_extensions=4.11.0=py39haa95532_0
tzdata=2024b=h04d1e81_0
ucrt=10.0.22621.0=h57928b3_1
urllib3=2.2.3=py39haa95532_0
utf8proc=2.6.1=h2bbff1b_1
vc=14.40=haa95532_2
vc14_runtime=14.42.34433=he29a5d6_23
vs2015_runtime=14.42.34433=hdffcdeb_23
wheel=0.44.0=py39haa95532_0
win_inet_pton=1.1.0=py39haa95532_0
xxhash=0.8.0=h2bbff1b_3
xz=5.4.6=h8cc25b3_1
yaml=0.2.5=he774522_0
yarl=1.18.0=py39h827c3e9_0
zlib=1.2.13=h8cc25b3_1
zstd=1.5.6=h8880b57_0

beautifulsoup4~=4.12.3
soupsieve~=2.6
pip~=24.2
attrs~=24.2.0
zlib~=1.2.13
wheel~=0.44.0
pysocks~=1.7.1
openssl~=3.4.0
pillow~=11.0.0
filelock~=3.13.1
dill~=0.3.8
future~=1.0.0
numpy~=1.26.4
pytz~=2024.1
requests~=2.32.3
colorama~=0.4.6
yaml~=0.2.5
pyyaml~=6.0.2
idna~=3.7
multidict~=6.1.0
propcache~=0.2.0
charset-normalizer~=3.3.2
setuptools~=75.1.0
cffi~=1.17.1
typing_extensions~=4.11.0
regex~=2024.9.11
fsspec~=2024.3.1
snappy~=1.2.1
aiohttp~=3.10.5
yarl~=1.18.0
pandas~=2.2.3
pyarrow~=16.1.0
markupsafe~=2.1.3
numexpr~=2.10.1
python-dateutil~=2.9.0post0
xxhash~=0.8.0
async-timeout~=4.0.3
aiosignal~=1.2.0
frozenlist~=1.5.0
aiohappyeyeballs~=2.4.3
certifi~=2024.8.30
tzdata~=2024b
urllib3~=2.2.3
huggingface_hub~=0.24.6
tqdm~=4.66.5
packaging~=24.1
multiprocess~=0.70.15
datasets~=2.19.1
six~=1.16.0
tokenizers~=0.21.0
safetensors~=0.4.5
lyricsgenius~=3.0.1
transformers~=4.47.0
accelerate~=1.2.0
jinja2~=3.1.4
torchvision~=0.20.1
torchaudio~=2.5.1
psutil~=6.1.0